// src/app/shared/model/documentsModel.ts
export interface Documents {
    documentId: Number;
    activityid: string;
    requestNo: string;
    moduleRefId: string;
    docType: string;
    docId: string;
    docSubject: string;
    docContent: string;
    docFileName: string;
    docContentType: string;
    objectUrl: string;
    objectId:string;
    docNatureId: string;
    ipAddress: string;
    subTime: string;
    createdBy:string;   
    
}